backend-plus-old-frontend/api/pyproject.toml
[project]
name = "webscraper"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "Flask-Cors>=5.0.0",
    "beautifulsoup4>=4.13.3",
    "boto3>=1.34.0",
    "botocore>=1.34.0",
    "cssbeautifier>=1.15.3",
    "dnspython>=2.7.0",
    "email-validator>=2.2.0",
    "flask-sqlalchemy>=3.1.1",
    "flask>=3.1.0",
    "geoip2>=5.0.1",
    "gunicorn>=23.0.0",
    "maxminddb>=2.6.3",
    "openai>=1.62.0",
    "psycopg2-binary>=2.9.10",
    "pyOpenSSL>=25.0.0",
    "pyright>=1.1.394",
    "python-dotenv>=1.0.0",
    "requests>=2.32.3",
    "sentry-sdk>=2.21.0",
    "toml>=0.10.2",
    "trafilatura>=2.0.0",
    "uuid7>=0.1.0",
]

[build-system]
requires = ["setuptools>=42.0", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
add-dep = "scripts.add_dep:main"

[tool.setuptools.package-data]
types_def = ["py.typed"]
utils = ["py.typed"]

[tool.setuptools.packages]
find = {where = ["."], exclude = ["static", "templates", "attached_assets"]}

[tool.pyright]
include = ["."]
exclude = [
    "**/node_modules",
    "**/__pycache__",
    "**/.venv/**",
    "**/site-packages/**",
    "/Library/Frameworks/Python.framework/**",
    "~/Library/**",
    "/Library/**",
    "/bin/**",
    "build",
    "dist"
]
ignore = [
    ".venv",
    "site-packages",
    "/Library/Frameworks/Python.framework",
    "/bin"
]
stubPath = "types_def/stubs"
reportMissingImports = true
reportMissingTypeStubs = false
pythonVersion = "3.12"
pythonPlatform = "Darwin"
typeCheckingMode = "basic"
useLibraryCodeForTypes = true

[tool.pylance]
reportMissingModuleSource = "warning"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
mypy_path = "types_def/stubs"

[[tool.pyright.executionEnvironments]]
root = "."
extraPaths = ["/Library/Frameworks/Python.framework/Versions/3.12/lib/python3.12/site-packages"]

